requirements.txt
selene>=2.0.0rc4
pytest>=7.4.2
Selenium>=4.12.0
pydantic
pydantic_settings
python-dotenv~=1.0.0
allure-pytest>=2.13.2
Appium-Python-Client>=3.0.0
pydantic-settings>=2.0.3
requests